+++ lib/native_admob_flutter.dart
@@ -1,6 +1,6 @@
 export 'src/native_ad_widget.dart';
 export 'src/native_ads.dart';
-export 'src/controller.dart';
+export 'src/controller/controller.dart';
 export 'src/utils.dart' hide colorExtension;
 
 export 'src/layout_builder/layout_builder.dart';