+++ lib/src/controller/controller.dart
@@ -3,9 +3,22 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
-import '../native_admob_flutter.dart';
+import '../../native_admob_flutter.dart';
+import 'options.dart';
 
-enum AdEvent { impression, clicked, loadFailed, loaded, loading, undefined }
+export 'options.dart';
+
+enum AdEvent {
+  impression,
+  clicked,
+  loadFailed,
+  loaded,
+  loading,
+  muted,
+  undefined,
+}
+
+enum AdVideoEvent { start, play, pause, end, mute }
 
 class NativeAdController {
   final _key = UniqueKey();
@@ -14,6 +27,9 @@ class NativeAdController {
   String get id => _key.toString();
 
   final _onEvent = StreamController<Map<AdEvent, dynamic>>.broadcast();
+
+  List<String> _muteThisAdReasons = [];
+  List<String> get muteThisAdReasons => _muteThisAdReasons;
 
   /// Listen to the events the controller throws
   ///
@@ -37,6 +53,12 @@ class NativeAdController {
   ///       break;
   ///     case AdEvent.clicked;
   ///       print('clicked');
+  ///       break;
+  ///     case AdEvent.muted:
+  ///       showDialog(
+  ///         ...,
+  ///         builder: (_) => AlertDialog(title: Text('Ad muted')),
+  ///       );
   ///       break
   ///     default:
   ///       break;
@@ -45,11 +67,17 @@ class NativeAdController {
   /// ```
   Stream<Map<AdEvent, dynamic>> get onEvent => _onEvent.stream;
 
+  final _onVideoEvent =
+      StreamController<Map<AdVideoEvent, dynamic>>.broadcast();
+  Stream<Map<AdVideoEvent, dynamic>> get onVideoEvent => _onVideoEvent.stream;
+
   /// Channel to communicate with plugin
   final _pluginChannel = const MethodChannel("native_admob_flutter");
 
   /// Channel to communicate with controller
   MethodChannel _channel;
+
+  bool _attached = false;
 
   /// Creates a new native ad controller
   NativeAdController() {
@@ -65,6 +93,15 @@ class NativeAdController {
     _pluginChannel.invokeMethod("initController", {"id": id});
   }
 
+  void attach() {
+    assert(
+      !_attached,
+      'This controller has already been attached to a native ad. You need one controller for each native ad.',
+    );
+    if (_attached) return;
+    _attached = true;
+  }
+
   /// Dispose the controller. Once disposed, the controller can not be used anymore
   ///
   /// Usage:
@@ -78,10 +115,31 @@ class NativeAdController {
   void dispose() {
     _pluginChannel.invokeMethod("disposeController", {"id": id});
     _onEvent.close();
+    _onVideoEvent.close();
   }
 
   /// Handle the messages the channel sends
-  Future<Null> _handleMessages(MethodCall call) async {
+  Future<void> _handleMessages(MethodCall call) async {
+    if (call.method.startsWith('onVideo')) {
+      switch (call.method) {
+        case 'onVideoStart':
+          _onVideoEvent.add({AdVideoEvent.start: null});
+          break;
+        case 'onVideoPlay':
+          _onVideoEvent.add({AdVideoEvent.play: null});
+          break;
+        case 'onVideoPause':
+          _onVideoEvent.add({AdVideoEvent.pause: null});
+          break;
+        case 'onVideoMute':
+          _onVideoEvent.add({AdVideoEvent.mute: null});
+          break;
+        case 'onVideoEnd':
+          _onVideoEvent.add({AdVideoEvent.end: null});
+          break;
+      }
+      return;
+    }
     switch (call.method) {
       case "loading":
         _onEvent.add({AdEvent.loading: null});
@@ -98,17 +156,26 @@ class NativeAdController {
       case "onAdImpression":
         _onEvent.add({AdEvent.impression: null});
         break;
+      case "onAdMuted":
+        _onEvent.add({AdEvent.muted: null});
+        break;
       case 'undefined':
       default:
         _onEvent.add({AdEvent.undefined: null});
         break;
+    }
+    if (call.arguments != null && call.arguments is Map) {
+      final Map args = call.arguments as Map;
+      if (args.containsKey('muteThisAdReasons')) {
+        _muteThisAdReasons = args['muteThisAdReasons'] ?? [];
+      }
     }
   }
 
   /// Load the ad.
   ///
   /// If [unitId] is not specified, uses [NativeAds.nativeAdUnitId]
-  void load([String unitId]) {
+  void load({String unitId, NativeAdOptions options}) {
     // assert(
     //   NativeAds.isInitialized,
     //   'You MUST initialize the ADMOB before requesting any ads',
@@ -117,12 +184,29 @@ class NativeAdController {
     assert(id != null, 'The ad unit id can NOT be null');
     _channel.invokeMethod('loadAd', {
       'unitId': id ?? NativeAds.testAdUnitId,
-      // "numberAds": numberAds,
+      'options': (options ?? NativeAdOptions()).toJson(),
     });
   }
 
   /// Request the UI to update when changes happen
   void requestAdUIUpdate(Map<String, dynamic> layout) {
-    _channel.invokeMethod('updateUI', {'layout': layout});
+    print('requested ui update');
+    _channel.invokeMethod('updateUI', {'layout': layout ?? {}});
   }
+
+  /// Mutes This Ad programmatically.
+  ///
+  /// Use null to Mute This Ad with default reason.
+  void muteAd([int reason]) {
+    // assert(reason != null);
+    // assert(
+    //   muteThisAdReasons.length > reason,
+    //   'The specified reason has not been found',
+    // );
+    _channel.invokeMethod('muteAd', {'reason': reason});
+  }
+}
+
+extension map<K, V> on Map<K, V> {
+  V get(K key) => this[key];
 }